Controllerの機能を分割
